+++ python/python/glide/__init__.py
@@ -24,6 +24,7 @@ from glide.async_commands.sorted_set import (
     RangeByLex,
     RangeByScore,
     ScoreBoundary,
+    ScoreModifier,
 )
 from glide.async_commands.transaction import ClusterTransaction, Transaction
 from glide.config import (
@@ -98,6 +99,7 @@ __all__ = [
     "RangeByIndex",
     "RangeByLex",
     "RangeByScore",
+    "ScoreModifier",
     "StreamAddOptions",
     "StreamTrimOptions",
     "TrimByMaxLen",