+++ python/python/glide/async_commands/transaction.py
@@ -22,6 +22,7 @@ from glide.async_commands.sorted_set import (
     RangeByLex,
     RangeByScore,
     ScoreBoundary,
+    ScoreModifier,
     _create_zrange_args,
     _create_zrangestore_args,
 )
@@ -1965,6 +1966,46 @@ class BaseTransaction:
         return self.append_command(
             RequestType.ZDiffStore, [destination, str(len(keys))] + keys
         )
+
+    def bzmpop(
+        self: TTransaction,
+        keys: List[str],
+        modifier: ScoreModifier,
+        timeout: float,
+        count: Optional[int] = None,
+    ) -> TTransaction:
+        """
+        Blocks the connection until it pops and returns a member-score pair from the first non-empty sorted set, with
+        the given keys being checked in the order they are provided. The optional `count` argument can be used to
+        specify the number of elements to pop, and is set to 1 by default. The number of popped elements is the minimum
+        from the sorted set's cardinality and `count`.
+
+        BZMPOP is the blocking variant of ZMPOP.
+
+        BZMPOP is a client blocking command, see https://github.com/aws/glide-for-redis/wiki/General-Concepts#blocking-commands for more details and best practices.
+
+        See https://valkey.io/commands/bzmpop for more details.
+
+        Args:
+            keys (List[str]): The keys of the sorted set.
+            modifier (ScoreModifier): The element pop criteria - either ScoreModifier.MIN or ScoreModifier.MAX to pop
+                members with the lowest/highest scores accordingly.
+            timeout (float): The number of seconds to wait for a blocking operation to complete. A value of 0 will
+                block indefinitely.
+            count (Optional[int]): The number of elements to pop.
+
+        Command response:
+            Optional[List[Union[str, Dict[str, float]]]]: A two-element list containing the key name of the set from
+                which elements were popped, and a member-score dict. If no members could be popped and the timeout
+                expired, returns None.
+
+        Since: Redis version 7.0.0.
+        """
+        args = [str(timeout), str(len(keys))] + keys + [modifier.value]
+        if count is not None:
+            args = args + ["COUNT", str(count)]
+
+        return self.append_command(RequestType.BZMPop, args)
 
     def dbsize(self: TTransaction) -> TTransaction:
         """